refactor: для вычисляемых свойств применен reselect
