+++ src/components/burger-ingredients-category/burger-ingredients-category.jsx
@@ -3,6 +3,7 @@ import PropTypes from "prop-types";
 
 import { useSelector } from "react-redux";
 
+import { selectCounters } from "../../services/slices/burger-constructor-slice";
 import { INGREDIENT_PROP_TYPES } from "../../utils/constants";
 import { TYPES_OF_INGREDIENTS } from "../../utils/constants";
 
@@ -11,16 +12,7 @@ import BurgerIngredient from "../burger-ingredient/burger-ingredient";
 import styles from "./burger-ingredients-category.module.scss";
 
 const BurgerIngredientsCategory = ({ type, list }) => {
-  const { buns, ingredients } = useSelector((state) => state.burgerConstructor);
-  const counters = {};
-
-  [...buns, ...ingredients].forEach((ingredient) => {
-    if (!counters[ingredient._id]) {
-      counters[ingredient._id] = 0;
-    }
-
-    counters[ingredient._id] += 1;
-  });
+  const counters = useSelector(selectCounters);
 
   return (
     <>
